[consensus] Improve test code for safety rules
Move to a library style of testing that enables us to reuse the same
test suite for all the variants of safety rules.
Add a simple end-to-end test that uses something more complex than
rounds to verify serialization a little more thoroughly.
Add local testing so we don't just leverage consensus code base for
testing.