[consensus] Add a separation layer between TCB and Consensus
I had hoped to use fork for this so we could get actual process
separation, but Rust does not support that. So for now, we'll leverage a
separate thread to evaluate the potential costs of separating the TCB.

This PR also introduces some interface changes that come as a result of
requiring mutability on the network stack.