[language] fix bugs in cost synthesis
It turned out the tests for cost-synthesis wasn't running in the way we
expected it to run and a few bugs sneaked in. This fixes some of them:
  1. Sub does not work with u8/u128.
  2. Lack of handling of overflow and division by zero.
  3. For each bytecode instruction, only the first iteration is executed.
  4. It still generates a few deprecated bytecode instructions.

Note this does not yet reenable the tests in a meaningful way. Some
other issues need to be fixed before this can be fully turned on again.