Introduce the ability to define native move procedures and helper functions in Boogie.
This adds the ability to add a file `<name>.prover.bpl` which lives SxS with `<name>.mvir` and contains
boogie definitions for native move functions as well as helper functions for the specifications.

The helper functions are a bit of a hack as we cannot type check them in the spec_translator, but they
should work as long as they result integers and other primitive values.

Existing generation of stubs for native functions is now behind a flag which is by default false.
The feature hasn't yet used to define a lot of helpers and/or override definitions for native
functions beyond what the compiler also generates, which happens in a separate step.

Closes: #2249
Approved by: shazqadeer