[vm] move derivation of type tags to views.rs
This function lives in the interpreter, but it is also useful elsewhere (e.g., the prover). In addition, some of the logic in the old implementation is a duplicate of functionality already available int the views.

This is a first step in a broader effort to put all of the useful procedures for converting back-and-forth between StructTag/TypeTag and the bytecode module types (StructHandle etc.) in one place.