Source level error reporting.
This implements source level error reporting, both for type checking errors, as well as Boogie errors.

Example of a verification checking error:

```
error:  A postcondition might not hold on this return path.
- libra_account.mvir:339:8
339 |         ensures global<Self.T>(txn_sender).balance.value == old(global<Self.T>(txn_sender).balance.value) - amount
help: Execution trace for previous error:
    at libra_account.mvir:352:8
    at libra_account.mvir:324:4
    at libra_account.mvir:342:15
    ...
```

As a side-effect, boogie output is not longer dumped to terminal. Instead it is written to a file `<out>.bpl.log`.

Details:

- All code generation is now directed via a new helper `CodeWriter` which tracks location information.
- The existing `SourceMap` functionality is used to populate the `CodeWriter` with source location info.
  It is also used to give locations to functions and other items in the environment for type check
  error reporting.
- A new helper `BoogieWrapper` is introduced to abstract Boogie execution, parsing of its output,
  and converting locations.
- Error reporting is handled via the well-known `codespan_reporting` crate.

The introduction of the `CodeWriter` caused a number of changes in the bytecode_translator, because
the design requires sequential writing of code instead of composing it via code snippet strings. This
lead to slight differences in code layout and many changes in golden files.