[language][prover] Added a test and fixed some bugs related to vectors.
1. Added basic tests test_mvir/verify-vector.mvir
2. Fixed an off-by-one error in src/prelude.bpl: RemoveValueArray
3. Added Boogie command-line option "-useArrayTheory" because Z3 array
extension we rely on need it. Otherwise, Boogie does not set invalid
elements of arrays to DefaultValue
4. Added a new variant ErrorType() to TypeValue and a new variant
Error() to Value, and added axioms to make the DefaultTypeValue=ErrorType()
and DefaultValue=Error().  The purpose of this change is to make it a little
easier to understand Boogie counter-models -- otherwise, Boogie assigns
random values that change with every change to the .bpl file, making it
much harder to see errors.
5. Added some short instructions for generating output.bpl in README.md.
I expect that this will be short-lived because there will be a better way.