[language] VM benchmark skeleton
This is a simple benchmark that compiles a file - `bench.move` - that
contains a `Bench` Module with functions that the bench will invoke and measure.
The move file has some vague reference around how long they run.
I checked that by adding more instructions the benchmark seems to move
accordingly so this may be a good playground to play with.