[CI] MIRAI fixes in CI
I've simplified how we obtain the base commit that the PR is working on top of +
I'm fetching the PR with a depth of 50 historical commits to make sure we can find the target base commit in there.

Also, some MIRAI runs have failed, and with the current way we run it we cannot see why on Github.
I replaced:

```
$ [command] > output
```

with:

```
$ [command] 2>&1 | tee ../mirai_results
```

which:

1. redirects stderr to stdout (`2>&1`)
2. prints and save the output to a file (via `tee`)