[language] Verify scripts (transactions) dependencies
this covers cross module dependencies verification when executing scripts.
Given how we changed transaction validation it is now possible to pass a completely
bogus script and though the script itself was verified, cross dependencies were not.
So it seems important to come up with a story soon and before we rework the loader and
the Libra/Move layers.
This code will go through a significant refactor/rework soon so this is a temporary solution.
In other words the implementation is not what we would like it to be but to make it
more proper it is too much change that does not seem appropriate given the rework.
We are also allowing e2e tests to be executing with any executor and not just one with white-lists.
Also added few related utilities here and there.
One behavior that is kind of "new" is that during execution we change every verification error
to a generic verification invariant violation. Verification error should never occur so that seems
a correct approach