Use local names provided by the bytecode-source-map crate where available.
This uses the actual assigned name (as provided by a let) instead of a name generated from the local index, where available. As we have transformed into a stackless machine, we have some unnamed locals, which will continue to use a generated name.

This is a prerequisite for getting model visualization working.

The implementation is based on the `bytecode-source-map` crate, which seems to do the right thing for our test cases out of the box(!).

This also renames the generated name `tN` to `__tN` to avoid clashes.

Closes: #2352
Approved by: tzakian