[move-prover] Prover model visualization in execution traces.
This adds variables to the generated boogie code which track position dependent information of the current value of a (mutable) boogie variable. The `boogie_wrapper.rs` is extended to extract values for those variables out of the model produced on verification failure and inject them in the execution trace.

This looks then for example like this (here we have injected a bug into LibraCoin.withdraw which doesn't actually subtract the withdrawn amount from the account. Therefore the amount in the `coin_ref` is not changed.)

```
error:  A postcondition might not hold on this return path.
- libra_coin.mvir:124:8
124 |         ensures coin_ref.value == old(coin_ref.value) - amount
help: Execution trace for previous error:
    at libra_coin.mvir:122:4: withdraw (entry)
        coin_ref = &LibraCoin.T{value = 3867},
        amount = 365
    at libra_coin.mvir:131:15: withdraw
        coin_ref = &LibraCoin.T{value = 3867},
        amount = 365,
        value = 3867
    at libra_coin.mvir:131:15: withdraw
        coin_ref = &LibraCoin.T{value = 3867},
        amount = 365,
        value = 3867
    at libra_coin.mvir:122:4: withdraw (exit)
        coin_ref = &LibraCoin.T{value = 3867},
        amount = 365,
        value = 3867,
        RET = LibraCoin.T{value = 365}
```

Here is another example of the test_empty2 method in `verify-vector.rs` where we inverted the post condition which illustrates how changes to variable state are tracked:

```
error:  A postcondition might not hold on this return path.
- verify-vector.mvir:16:4
16 |     ensures RET(0) != RET(1)
help: Execution trace for previous error:
    at verify-vector.mvir:15:4: test_empty2 (entry)
    at verify-vector.mvir:21:7: test_empty2
        ev1 = Vector.T{},
    at verify-vector.mvir:22:7: test_empty2
        ev1 = Vector.T{},
        ev2 = Vector.T{},
    at verify-vector.mvir:23:1: test_empty2
        ev1 = Vector.T{1},                    // <-- at this point we pushed 1
        ev2 = Vector.T{},
    at verify-vector.mvir:24:5: test_empty2
        ev1 = Vector.T{},                     // <-- at this point we have popped it
        ev2 = Vector.T{},
        x = 1,
    at verify-vector.mvir:15:4: test_empty2 (exit)
        ev1 = Vector.T{},
        ev2 = Vector.T{},
        x = 1,
        RET(0) = Vector.T{},
        RET(1) = Vector.T{}
```

The feature isn't fully stable yet, and bugs and imprecision are to be expected. Specifically, it appears that there is either a bug in the implementation or in Boogie/Z3 maps which lets some variables appear earlier in the trace as they are initialized.

The design introduces for each named variable a 'debug' variable which is a map `[Position]Value` in boogie, and which is updated whenever the variable is updated. Example:

```
var ev1: Value;
...
var debug#VerifyVector#test_empty2#0#ev1: [Position]Value;
...
call __t3 := Vector_empty(IntegerType());
__m := UpdateLocal(__m, __frame + 0, __t3);
debug#VerifyVector#test_empty2#0#ev1 := debug#VerifyVector#test_empty2#0#ev1[Position(440) := __t3];
...
```

Above, 440 is the byte index into the corresponding source.

Reference values are dealt with by storing the `Dereference(__m, r)` at each point it is being updated.

Care has to be taken for when a local is borrowed. To this end, the current algorithm approximates the result by first analyzing which variables are borrowed in a function, and then updating all the debug variables of all borrowed variables whenever a function with &mut parameters is called. This can likely be done more efficient by actually tracking which locals references can point to.

Global state isn't dealt with yet. One possible design is to look at `acquires` clauses from functions and track the state of the acquired resource type. One problem with this is that the global memory map is unbounded, so we can iterate its domain.

The values of the debug map variables are then extracted from the boogie model and injected into the execution trace, using a bunch of tricky/complicated (i.e. not yet ideally designed) code. This includes translating vectors and structs back into Move style representation.

Note that the implementation currently makes a lot of assumptions how the boogie model looks like and likely breaks if we use array theories or CVC4. However, the code should be robust -- it shouldn't crash but visualization may stop working.