[CI] removing MIRAI from CI
It looks like Github actions are limited when dealing with PRs from forked repositories.
The problem is: the `GITHUB_TOKEN` used to run the github action is the one from the PR repository, not the one from the main repository (libra/libra).
This makes it impossible for our github action to write a comment on the PR when MIRAI finds issues.
The fact that this is added as a check also confuses people (because MIRAI still tends to fail).

For this reason I'll remove this for now.