[move-prover] Improving model visualization.
The current approach had issues with some bogus values which where printed at code locations.  This introduces a new approach to track debug information which is more lightweight and flexible, and has less issues with correctness (e.g. in the presence of loops, and printing garbage values).

Until now, we have used special variables to track debug information. However, this causes problems if we assign to those variables or make assumptions over them:

1. If we assign, it must be a single assignment over the whole code body. Boogie generates different versions x@1, x@2, etc. for each update of a variable, but we can't see from the model which is the most recent version of x (x@1?, x@2?). This produced the bogus values.

2. If we assume (e.g. `assume debug$var$at_128 == v`) we run into issues with loops. If a code point is visited more than once during execution, we may get logical inconsistencies.

The new approach instead uses a global boolean function and makes an assumption over it. For instance, to track a variable update, we use function $DebugTrackLocal and use it e.g. as such:

```
  <<update var x>>
  assume $DebugTrackLocal(module_idx, func_idx, var_idx, location, value);
```

As it happens Boogie/Z3 stores all those assumes (and only those which have been reached) in a handy function table as such:

```
$DebugTrackLocal -> {
  0 4 0 855 (Integer 323) -> true
  0 4 1 855 (Integer 340282366920938463463374607431768211134) -> true
  0 5 0 1016 (Integer 323) -> true
  0 5 1 1016 (Integer 340282366920938463463374607431768211134) -> true
  else -> true
}
```

This is not only easier to analyze as the previous approach, but also more efficient as we do not need to introduce any special variables for debugging. It also is more flexible as other kind of information can be easily tracked. As an example, this PR also starts tracking abort positions:
```
  if (__abort_flag) { assume $DebugTrackAbort(module_idx, func_idx, location); goto Label_Abort; }
```

Besides refactoring the debug variable approach, this also cleans up the location handling. The problem here was that boogie execution traces do not necessarily point to the locations where updates happen. A new heuristic prints all variables until a certain point which haven't been shown before. This likely will need further refinment, specifically if comes to loops.