Improving model visualization.
It turns out that the problems in model visualization (variables appear with values they don't have) are a result of a basic restriction of what we can extract from a model. If we have a stareful variable which is updated multiple times, we cannot know what its 'current' value is. Example: consider

```
   var debug_var: [Position]Value;
   debug_var := EmptyPositionMap;
   ...
   debug_var[Position(33)] := v1;
   goto L;
   debug_var[Position(44)] := v2;
L: ...
```

Here we will have `debug_var@0`, `debug_var@1` and `debug_var@2` in the model, corresponding to the initialization and the two updates. But `debug_var@2` will have an arbitray value assigned even though the update was never executed. The model doesn't give as any information about the latest active version of a variable. This essentially renders us unable to reason about the current value of `debug_var` at a given execution point.

To solve this, the design for debug variables had to be changed (and is now very similar to Clark's original suggestion). Now we generate a single variable for each location, which is assigned to exactly once. Morover, we use `assume` to set the value, because assignments to unused variables are eliminated by boogie. Example reads now as:

```
   var debug_var#33: Value;
   var debug_var#44: Value;
   ...
   assume debug_var#33 == v1;
   goto L;
   assume debug_var#44 == v2;
L: ...
```

Luckily, boogie/z3 will not assign anything to `debug_var#44` if the `assume` is not executed. (A bit of a fragile state, because this is based on observation only).

Besides refactoring the debug variable approach, this also cleans up the location handling. The problem here was that boogie execution traces do not necessarily point to the locations where updates happen. A new heuristic prints all variables until a certain point which haven't been shown before. This likely will need further refinment, specifically if comes to loops.