[docker] Add config-builder to safety-rules image
safety-rules currently uses the same config as the validator, so we need
to be able to generate the config. Add config-builder binary to the
image so this is possible.

We also need to be able to pass a template config file to the validator.
The run script accepts a file path, but this isn't useful if the file
isn't there. Change to accept the content and write to file.