[CI] Run cluster test bench at land time
Run the test bench suite of cluster test at land time. This relies on
bors to push the PR to the auto branch. A self-hosted GitHub Actions
runner then picks up the push request and carrys out the cluster test
workflow. This workflow will block a PR from landing if the cluster test
fails. Currently, the workflow completes in 20 minutes.