[consensus] Remove some unwraps
These unwraps are unsafe once we move to a proper safety rules service
and aren't needed. I also removed some asserts as they really aren't
providing value outside of ensuring that we have a consistent
implementation. Do we really need to verify that a round obtained by
directly from a struct is the same round as acquired by a helper
function?