[consensus-types] Remove dependency on networking
This removes consensus-types on networking by:
* Removing proto definitions from each type
* Changing the ConsensusMsg to contain only bytes
* Introducing a LCS enum to replace the now missing ConsensusMsg variants

Most of this commit is boilerplate code removal, here's how it was not:
* An LCS enum does not allow to easily convert between types
(ProposalMsg -> ProposalUncheckedSignatures), while it is possible, it
ends up with some ugly serde macro stuff and the LCS enum takes in a
ProposalUncheckedSignatures.
* The network still depends on a proto, so completely eliminating proto
on outgoing messages is impossible without more code refactor, hence
proto was left also on the input functions.

At this point the consensus networking code does not really do much and
the integration tests verify functionality that should already be
evaluated at a lower layer. It might be worth considering to remove the
tests.