[language][move] Use a new "fun" keyword to for function definitions.
The previous syntax, where function definitions can begin with the
function name identifier, makes it hard to provide robust error messages
in some cases. In particular, this came up with the "address" directive,
which we prefer not to make into a keyword, when the account address is
not valid. We may make further changes to how we handle "address", but
adding a "fun" keyword also improves readability of Move code by making
it more obvious that something is a function definition. We intentionally
want to use "fun" instead of "fn" so that Move code looks distinct from
Rust code.

Closes: #2505
Approved by: tnowacki