[types] Ignore stale prefix when verifying ValidatorChangeProof
With this change, we now skip any stale ledger infos in the
prefix of a `ValidatorChangeProof`. It's useful to skip these
stale ledger infos to better support concurrent client requests.

For example, suppose the following:

1. My current trusted state is at epoch 5.
2. I make two concurrent requests to two validators A and B, who
   live at epochs 9 and 11 respectively.

If A's response returns first, I will ratchet my trusted state
to epoch 9. When B's response returns, I will still be able to
ratchet forward to 11 even though B's ValidatorChangeProof
includes a bunch of stale ledger infos (for epochs 5, 6, 7, 8).

Of course, if B's response returns first, we will reject A's
response as it's completely stale after ratcheting to epoch 11.