[RFC][language] Binary format changes
A review of the changes in the binary format.
Comments and code should be enough for us to have a conversation.

I want to point out I vote against all the names in this PR
so if anybody could come up with a palatable naming story that
would be great.

Should be - on paper - a good compromise between binary representation
and what we need to express for verification and execution.

good