[Storage] start_storage_service waits for server to start listening
In the tests we query storage with a client immediately after calling
this function. Sometimes the server does not seem to be available by the
time the client calls. The reason seems to be that the server creation
is async and the function might return before the server starts
listening.

As a workaround, try to connect to the server a couple of times in this
function. Panic if the server fails to respond in a few seconds.