[network] Remove nearly all proto from HealthChecker interface
This commit removes all but the HealthCheckerMsg proto and introduces
an LCS layer instead. This matches the earlier work done in https://github.com/libra/libra/pull/2502

Note: this does not introduce the structs for ping and pong as they are
just u32, but it does somewhat hurt the readability of the code. Not
sure if that makes it worth it to have the extra embedding. Another
alternative is to relabel the u32 as a proper type like a
HealthCheckerNonce or just nonce. Again not sure if that is worth it, so
punting on feedback.