[vm-runtime] move gas loading code to Libra VM
In a world where the Move VM is independent of the Libra VM, it would not be
aware of the exact details of how gas is loaded.

This is probably not going to be the final resting place for the gas
schedule fetching code, btw -- need to think about how to do this in a
principled way after we're further through the separation.