This PR integrates specification constructs into the Move compiler, up to the point of its expansion phase. As previously discussed, it implements the strategy to integrate Move language and specification syntactically, for consistency, but let them be separate for name resolution and type checking for agility.
The spec language has additional expression forms and at the same time, does not support all those forms the Move language does. In this PR we handle the new forms. A subsequent PR will translate the AST resulting from expansion into a more restricted AST which will be input to the Move prover.

Below is a list of the additional constructs which have been added to the Move parser:

-   a new construct called a specification block (`spec <Target> { ... }`) has been introduced. This contains specification elements like `aborts_if`, `ensures`, `invariant`, ghost variables, and specification helper functions. Spec blocks are currently independent members of a module and are associcated with items of the move program by referring via name to them.
-   Indexing and slicing for vectors (`e[i]` and `e[i..j]`). This required introduction of a new Exp AST form `Index`, as well as a new binary binary operator `..`.
-   Logical implication operator (`p ==> q`)
-   Lambda for quantification (`|x,y| x + y`)
-   Function types for specification helper functions (`define f(p: |u64|bool) { ... }`)

Extensions like the above are handled by the parser, but rejected during expansion phase if used outside of the `spec` context. Tests have been added which ensure that none of the extended expression constructs reach later phases of the move compiler. Some positive parsing test cases for the spec language have been added as well, but full testing will happen outside of the move-lang crate in a future PR, at the point we consume the specification language AST.