[types] add `TrustedState` for ratcheting latest verified state
+ `TrustedState` keeps track of our latest trusted state, including
the latest verified version and the latest verified validator set or
epoch change waypoint.

+ `TrustedState` ratchets this latest trusted state forward when
verifying a `UpdateToLatestLedgerRequest/Response` pair. In other words,
it keeps track of our current epoch, lets us safely transition to new
epochs, and ensures that we never go backwards to a previous epoch or
version.

+ Similar logic exists in the current libra client code; since onchain
didiscovery needs this logic as well, this common logic can be extracted
into a single place.