[types] Ensure we always use VerifierType to check ledger info
+ By consistently using VerifierType, we ensure that the ledger info
we're checking is actually in the next epoch and not just signed by the
current validator set (which could be the same in different epochs).

+ Without this fix, the test case `test_ratchet_fails_with_gap_in_proof`
in another PR #2700 fails. I will leave the regression test to that PR.

+ Are the any other properties we should be checking? It looks like all
we're currently check (this diff included) is the epoch number and the
validators' signatures. Should we check that the version number is
monotonically increasing? That the timestamp is monotonically increasing?

+ Our verification infrastructure is pretty convoluted at this point. We
should think about doing some further refactoring to make this less
error prone.

Closes: #2731
Approved by: zekun000