[logger] rewrite logger
This removes the slog and its dependencies for a substantially simpler
logger that depends on env_logger, log, and regex to accomplish the same
thing. This is a reduction of several crates from the TCB. This also
introduces a cleaner way for instantiating the global logger so all code
has been unified towards that path.

This reduces the total number of crates in the logger from 77 down to 10.

Will use cluster test to evaluate any additional impact.