[network] Add structs for new network wire protocol
GitHub Issue: https://github.com/libra/libra/issues/2878

Implementation:
We divide the wire spec directory structure into two subdirectories:
1. handshake - for the LibraNet handshake protocol. This contains the
   structs exchanged during the handshake phase, before any messages are
   sent.
2. messaging - for the LibraNet messaging protocol. This contains
   structs exchanged after handshake has been performed.

For the handshake protocol, we need to send the remote peer a bit vector
signalling the application protocols we support over certain LibraNet
messaging protocol version. To represent the bit vector, we create a new
BitVec struct which wraps a `Vec<u8>`, where each each bit of a u8 is
set to 0 or 1 depending on whether that position is set or not.