[move-prover] quantification over ranges
Filled in missing support for quantification over ranges in the
specification language.

There were some issues in generated boogie code with Value and int
types getting mixed up.

This caused Boogie to get type errors, which were typically reported
by the Move Prover like this:

bug: output.bpl(1208,438): Error: invalid type for argument 1 in application of IsEqual: int (expected: Value)

To address this, I added to the prelude
$vlen_value, which returns the length of a vector as a Value, not as
an int, and $select_vector_by_value, which indexes into a vector
using a Value (Integer) for the index instead of an int.
I am not yet confident that we have nailed all such problems, or that
my fixes are consistent.

I changed spec_translator.rs:translate_call to generate calls to these
functions for "len" and "index" nodes.

There was also an issue with nested quantified expressions.  The
generated Boogie code used constant names for the range and quantified
variable which clashed when there was a nested quantified expression.
I added a new function fresh_var_name(&str) that generates a string
for a unique variable name, and used those names in the generated
Boogie code.

   [move-prover]

    Filled in missing support for quantification over ranges in the
    specification language.

    There were some issues in generated boogie code with Value and int
    types getting mixed up.

    This caused Boogie to get type errors, which were typically reported
    by the Move Prover like this:

    bug: output.bpl(1208,438): Error: invalid type for argument 1 in application of IsEqual: int (expected: Value)

    To address this, I added to the prelude
    $vlen_value, which returns the length of a vector as a Value, not as
    an int, and $select_vector_by_value, which indexes into a vector
    using a Value (Integer) for the index instead of an int.
    I am not yet confident that we have nailed all such problems, or that
    my fixes are consistent.

    I changed spec_translator.rs:translate_call to generate calls to these
    functions for "len" and "index" nodes.

    There was also an issue with nested quantified expressions.  The
    generated Boogie code used constant names for the range and quantified
    variable which clashed when there was a nested quantified expression.
    I added a new function fresh_var_name(&str) that generates a string
    for a unique variable name, and used those names in the generated
    Boogie code.