[serde reflection] New tracer based on deserialization
Tracing Serde serialization (1) to produce an IDL-like format works well but it requires user-provided Rust values and does not guarantee an exhaustive result for enums.

This commit introduces the possibility to trace Serde deserialization (2). This approach guarantees exhaustivity but it may crash if the analysis generates an invalid value for a user-defined type.

Therefore, this commit also make it possible to mix both approaches: values provided by the user during (1) are recorded, and used during (2) to pass user-defined checks.