[consensus] implement LeaderReputation
We introduce a committed history based leader election implementation in this pr,
the heuristic is simple - if a author is active in the past a few
blocks, it gets higher chance to be elected again.

In case validators don't agree on the committed history, they'll timeout
and trigger a synchronization across board.

This PR only lays foundation of the impl without integrating with the
rest of consensus.