[move-prover] Slaying mivr based Move prover.
This removes all code in the old version of the prover in bytecode-to-boogie. The data analysis framework has been moved to stackless-bytecode-generator.

The only thing preserved for now are the MVIR test sources in `bytecode-to-boogie/test_mvir`. We keep them around until we are sure all tests are ported. These files aren't connected to any Rust test.