[key-manager] Add an interface for querying Libra
This adds the basic interfaces for reading from storage the current
state of the validator. Tests have been added to ensure this works
correclty. Because this is becoming a lot of code, this serves as a
checkpoint into the system build. Currently we can setup a secure
storage, ensure that the validator config and info are consistent with
it, and that the last rotaton happened recently. There's a lot of work
remaining.