[consensus] Improve test code by using correct runtime
Move test execution inside the tokio executor, and switch to using
tokio::time::timeout to control test timeouts. This also removes the crash
handler from the harness, as it was force killing the test process leading to
strange behavior on failed tests.

Instances of `async move` were changed to `async`, as the runtime cannot be
dropped from inside async code, as it calls `block_on` during Drop.

Note that this does not eliminate futures::executor from the codebase.