[network] Remove default Yamux upgrade in transport
With this change, connections are no longer upgraded to Yamux by
default. Instead, Peer actor receives raw socket from PeerManager and
then upgrades it to yamux. This means that Peer is now the only place in
the code that uses the notion of substream. Some yamux references also
remain in PeerManager tests.