[State-sync] Fix a bug in the request epoch calculation
Summary:
During the request processing a state-sync sends an optimistic request for the next chunk.
The epoch number of that request should be either current epoch, or the next epoch if the received chunk is the last one in its current epoch.
We had a bug, in which the condition of "last chunk in its epoch" was checked by inspecting the `next_validator_set` of the received ledger info.
Even if the received ledger info is the last one in its epoch, the received chunk might still be in the middle. As a result, we could erroneously ask for an epoch that we cannot verify, which broke the download pipeline, and state sync would become very slow.

The current change is checking for the "last chunk in its epoch" by both inspecting the `next_validator_set` and by comparing the future would-be version with the version of the ledger info.

Tests: unit tests

Closes: #3130
Approved by: phoenix-antigravity