[LBT][GHA] combine build and test jobs
LBT divides the workflow into has two jobs: build and test. Only the
test job will comment on the PR when it fails.  If the build job fails,
the PR owner is left in the dark, pondering why their PR isn't landing.

This commit is a temp fix to combine the jobs to leverage the comment
step.  The proper long term fix, however, is to generalize this comment
step into a standalone action so that it can be reused in other
jobs or workflows.

Closes: #3152
Approved by: andll