[move-prover] Making update invariants on nested data structures work.
Update invariants (and pack/unpack) on nested data structures haven't been correctly implemented before. This PR fixes some holes to this end. There is still an open problems with Vector<StructWithInvariant> which need to be fixed in a subsequent PR.

This also repairs field ordering in pack instructions. Shaz fixed this a couple of days ago, but it was accidently reverted by the large file format PR which went in afterwards. We did not had a test wiht multiple fields to catch this. This PR also indirectly adds test for this use case.

Closes: #3268
Approved by: shazqadeer